Unai

+++ mavenproject1/src/main/java/daos/MysqlvideojuegoDAO.java
@@ -38,11 +38,6 @@ public class MysqlvideojuegoDAO implements VideojuegoDAO {
             stmt.setInt(5, videojuego.getTotal_sessions());
             stmt.setDate(6, videojuego.getLast_session() );
             stmt.executeUpdate();
-
-//            ResultSet keys = stmt.getGeneratedKeys();
-//            if (keys.next()) {
-//                videojuego.setGame_id(keys.getInt(1));
-//            }
         }
     }
 
@@ -63,12 +58,19 @@ public class MysqlvideojuegoDAO implements VideojuegoDAO {
 }
 
     @Override
-public void deleteVideojuego(int gameId) throws SQLException {
-    String sql = "DELETE FROM videojuegos WHERE game_id = ?";
+public boolean deleteVideojuego(int gameId) throws SQLException {
+    Videojuego eliminado= getVideojuego(gameId);
+        if (eliminado==null) {
+            return false;
+        }else{
+            String sql = "DELETE FROM videojuegos WHERE game_id = ?";
     try (PreparedStatement stmt = connection.prepareStatement(sql)) {
         stmt.setInt(1, gameId);
         stmt.executeUpdate();
     }
+            return true;
+        }
+    
 }
 
 @Override
